refactor: 딜러의 승패 판단을 분리
