+++ docs/README.md
@@ -12,6 +12,7 @@
 - 플레이어 게임 결과(PlayerGameResult)
 - 돈(Money)
 - 배팅정보(Bettings)
+- 승패 판단(Judge)
 
 ## 블랙잭 승패 규칙
 
@@ -99,3 +100,7 @@
 - [x] 배팅 금액은 1000원 단위여야 한다.
 - [x] 딜러 수익을 알 수 있다.
 - [x] 플레이어들의 수익을 알 수 있다.
+
+### 승패 판단
+
+- [x] 플레이어와 딜러를 통해 플레이어의 승패를 판단할 수 있다.
