+++ src/main/java/blackjack/domain/participant/Dealer.java
@@ -1,9 +1,8 @@
 package blackjack.domain.participant;
 
+import blackjack.domain.Judge;
 import blackjack.domain.PlayerGameResult;
 import blackjack.domain.card.Card;
-import blackjack.domain.participant.Participant;
-import blackjack.domain.participant.Player;
 import java.util.List;
 
 public class Dealer extends Participant {
@@ -12,43 +11,15 @@ public class Dealer extends Participant {
     private static final int STAND_BOUND = 17;
     private static final int FIRST_VISIBLE_CARD_INDEX = 1;
 
+    private final Judge judge;
+
     public Dealer() {
         super(DEALER_NAME);
+        this.judge = new Judge();
     }
 
     public PlayerGameResult judge(Player player) {
-        int playerScore = player.calculateScore();
-        int dealerScore = calculateScore();
-
-        if (player.isBust()) {
-            return PlayerGameResult.LOSE;
-        }
-
-        if (player.isBlackJack()) {
-            return judgeWhenPlayerIsBlackjack();
-        }
-
-        return judgeWhenPlayerNormalScore(playerScore, dealerScore);
-    }
-
-    private PlayerGameResult judgeWhenPlayerIsBlackjack() {
-        if (isBlackJack()) {
-            return PlayerGameResult.PUSH;
-        }
-
-        return PlayerGameResult.BLACKJACK_WIN;
-    }
-
-    private PlayerGameResult judgeWhenPlayerNormalScore(int playerScore, int dealerScore) {
-        if (isBust() || playerScore > dealerScore) {
-            return PlayerGameResult.WIN;
-        }
-
-        if (isBlackJack() || playerScore < dealerScore) {
-            return PlayerGameResult.LOSE;
-        }
-
-        return PlayerGameResult.PUSH;
+        return judge.compare(this, player);
     }
 
     public List<Card> getVisibleCards() {
